Settingup the instanceID in the datafetcher instead of the App class
