+++ App/Services/Fetcher.cs
@@ -8,6 +8,9 @@ using System.Collections.ObjectModel;
 using Plugin.FirebasePushNotifications;
 using System.Net.Sockets;
 
+#if ANDROID
+using static Android.Provider.Settings;
+#endif
 
 #if IOS
 using Maui.RevenueCat.InAppBilling.Services;
@@ -23,6 +26,8 @@ public class Fetcher
     public static string ProdHost { get; } = "api.gamhub.io";
     private static string _dateFormat = "dd-MM-yyy_HH:mm:ss";
     private Session CurrentSession { get; set; }
+
+    public string InstanceID { get; set; }
     public Service WebService { get; private set; }
     public static Collection<Source> Sources { get; set; }
 
@@ -1062,7 +1067,7 @@ public class Fetcher
             if (string.IsNullOrEmpty(instanceID))
             {
                 // Reissue an instanceID
-                instanceID = await (App.Current as App).SetupInstance();
+                instanceID = await SetupInstance();
             }
     #if DEBUG
             Debug.WriteLine($"ApiKey: {apiKey}");
@@ -1087,6 +1092,40 @@ public class Fetcher
     }
 
     /// <summary>
+    /// Setup the instance for a device
+    /// </summary>
+    /// <returns>a task returning the id of the instance</returns>
+    public async Task<string> SetupInstance()
+    {
+
+#if IOS
+        string instanceID = await SecureStorage.Default.GetAsync(AppConstant.InstanceIdKey);
+        if (string.IsNullOrEmpty(instanceID))
+        {
+            await SecureStorage.Default.SetAsync(AppConstant.InstanceIdKey, instanceID = UIKit.UIDevice.CurrentDevice.IdentifierForVendor.ToString().ToLower().Replace("-", string.Empty).Substring(0,30));
+
+        }
+
+#if DEBUG
+        Debug.WriteLine($"Instance: {instanceID}");
+#endif
+        return
+#endif
+        InstanceID =
+#if ANDROID
+            Secure.GetString(Android.App.Application.Context.ContentResolver, Secure.AndroidId);
+#if DEBUG
+        Debug.WriteLine($"Instance: {InstanceID}");
+#endif
+        await SecureStorage.Default.SetAsync(AppConstant.InstanceIdKey, InstanceID);
+        return InstanceID;
+#elif IOS
+            instanceID;
+#endif
+    }
+
+
+        /// <summary>
     /// Adding a hook to an article
     /// </summary>
     /// <param name="article">article hooked</param>
